CODA-125 - Customisable navigation
