+++ README.md
@@ -6,6 +6,26 @@ A simple GO based CMS application.
 ### Motivation
 Goal of this project is to create a flexible web-application CMS. 
 
+### Configuration
+WebApp is a configurable application. Configuration can be made by overwriting the input file `webapp-config.json`.
+The config inside is a JSON object with the following properties that can be customised: 
+
+#### Navigation
+Key: `"navigation"` - array of objects
+```json
+{
+  "id": "navigation-id", // string
+  "label": "Label to display", // string
+  "href": "/path-to-navigate", // string
+  "isRoot": false, // bool - is item visible after logging-in
+  "children": [{
+    "id": "sub-navigation-d",
+    "label": "Sub label to display",
+    "href": "/path-to-navigate/sub"
+  }] // array of objects - only one level of nesting
+ } 
+```
+
 ### Local development
 #### To run locally
 1. run `make run-services` in one terminal
